fix: Advanced tab – single useEffect for guaranteed DOM order
Replace the mixed React-component + Obsidian-API pattern in
ServerActionsSection (which caused AdvancedEnvSection to render
before 'Server configuration' due to effect scheduling) with a
single useRef/useEffect block that creates all three sections
sequentially in the Obsidian DOM API:

  1. Server configuration  (Review .env & restart)
  2. Advanced configuration (total control)  ← details/summary + env textarea
  3. Reprocess failed documents

This is the same pattern NeuralSection uses and guarantees the
correct visual order. Remove AdvancedEnvSection.tsx (now inlined).
